requirements.txt
kafka-python==2.0.1
psycopg2==2.7.7
PyYAML==5.3.1
